+++ iris-thaumantias/src/utils/index.ts
@@ -3,3 +3,4 @@ export * from './iconDefinitions';
 export * from './aiExtensionsBlocklist';
 export * from './plantUmlProcessor';
 export * from './uncommittedFilesCollector';
+export * from './gitStatusChecker';